tools/shared/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "openclaw-shared"
version = "0.1.0"
description = "Shared utilities for OpenClaw industry tools"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "python-dateutil>=2.9",
    "httpx>=0.27",
    "apscheduler>=3.10",
]

[project.optional-dependencies]
mlx = ["mlx-lm>=0.21"]
messaging = ["twilio>=9.0", "python-telegram-bot>=21.0"]
all = ["openclaw-shared[mlx,messaging]"]

[tool.setuptools.packages.find]
include = ["openclaw_shared*"]
